Detail fixes

+++ common/common_utls/src/main/java/com/example/common_utls/NavigationRoute.kt
@@ -3,7 +3,7 @@ package com.example.common_utls
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.navigation.NavHostController
 
-sealed class NavigationRoute(val route: String, val name: String? = null, val id: Int? = null) {
+sealed class NavigationRoute(val route: String, val name: String? = null, val id: Long? = null) {
     object MoviesScreen : NavigationRoute(route = "movies_screen")
     object DetailScreen : NavigationRoute(route = "detail_screen/{id}")
     object SearchScreen : NavigationRoute(route = "search_screen")
