+++ details/details_data/src/main/java/com/example/details_data/model/BelongsToCollection.kt
@@ -2,7 +2,7 @@ package com.example.details_data.model
 
 data class BelongsToCollection(
     val backdrop_path: String?,
-    val id: Int?,
+    val id: Long?,
     val name: String?,
     val poster_path: String?
 )
