+++ details/details_data/src/main/java/com/example/details_data/network/DetailApi.kt
@@ -6,5 +6,5 @@ import retrofit2.http.Path
 
 interface DetailApi {
     @GET("/3/movie/{id}")
-    suspend fun getDetail(@Path("id") id: Int): DetailResponse
+    suspend fun getDetail(@Path("id") id: Long): DetailResponse
 }
