+++ details/details_domain/src/main/java/com/example/details_domain/repository/DetailRepository.kt
@@ -3,5 +3,5 @@ package com.example.details_domain.repository
 import com.example.details_domain.model.Detail
 
 interface DetailRepository {
-    suspend fun getDetail(id: Int): Detail
+    suspend fun getDetail(id: Long): Detail
 }
